Refactored ItemListAdapter and MyAdapter into one ProductAdapter class
